Extract parsePeers()
